feat: reduce candiates over substituter instead of an inefficient map
substituters are reduced over in order of their priority.
the previous map ended up behaving like a sequencial query instead of
a parallel one on big closures, so by reducing instead of mapping we
ensure not to incur in unecessary query work.
priorities should be set in such a way that the "closest" cache is also
set to the lowest priority, e.g. 1